[3.4] Integration test setup
Updated circleci to use new rust image
Set up skeleton for integration test
