+++ Cargo.toml
@@ -3,6 +3,20 @@ name = "zero2prod"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+# We could use any path here, but we are foloowing the community convention
+# We could specify a library name using the 'name' field. If unspecified,
+# cargo will default to 'package.name', which is what we want
+path = "src/lib.rs"
+
+# Notice the double square brackets: it's an array in TOML's syntax.
+# We can only have one library in a project, but we can have multiple binaries!
+# If you want to manage multiple libraries in the same repository
+# have a look at the workspace feature - we'll cover it later on
+[[bin]]
+path = "src/main.rs"
+name = "zero2prod"
+
 [target.x86_64-apple-darwin]
 rustflags = ["-C", "link-arg=-fuse-ld=/usr/local/bin/zld"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
