+++ tests/health_check.rs
@@ -1,0 +1,6 @@
+use zero2prod::run;
+
+#[test]
+fn dummy_test() {
+    run();
+}
